requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
Django==2.1.1
django-cors-headers==2.4.0
djangorestframework==3.8.2
Fabric3==1.14.post1
idna==2.7
paramiko==2.4.1
pyasn1==0.4.4
pycparser==2.18
PyNaCl==1.2.1
pytz==2018.5
raven==6.9.0
six==1.11.0
